requirements.txt
Django==2.1.2
mysqlclient==1.3.13
pytz==2018.7
flake8==3.6.0
Jinja2==2.10
graphene-django==2.2.0
django-graphql-jwt==0.2.0
